+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -440,6 +440,7 @@ DeploymentsProvider = Literal[
     "litellm",
     "openailike",
     "cerebras",
+    "bytedance",
 ]
 
 DeploymentsRole = Literal[