+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -367,6 +367,7 @@ ListPromptVersionsProvider = Literal[
     "litellm",
     "openailike",
     "cerebras",
+    "bytedance",
 ]
 
 ListPromptVersionsRole = Literal[