+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -179,7 +179,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-15T12:52:59.508Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-25T14:18:53.694Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")